create api simple for timesheet and employee

+++ src/main/java/com/example/coffee_shop_chain_management/service/EmployeeService.java
@@ -5,6 +5,8 @@ import com.example.coffee_shop_chain_management.entity.Branch;
 import com.example.coffee_shop_chain_management.entity.Employee;
 import com.example.coffee_shop_chain_management.repository.BranchRepository;
 import com.example.coffee_shop_chain_management.repository.EmployeeRepository;
+import com.example.coffee_shop_chain_management.response.APIResponse;
+import com.example.coffee_shop_chain_management.response.EmployeeResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,9 +19,24 @@ public class EmployeeService {
     @Autowired
     private BranchRepository branchRepository;
 
-    public List<Employee> getAllEmployees() {
-        return employeeRepository.findAll();
+    public APIResponse<List<EmployeeResponse>> getAllEmployees() {
+        List<Employee> employees = employeeRepository.findAll();
+        return new APIResponse<>(employees.stream().map(this::toEmployeeResponse).toList(), "Employees retrieved successfully", true);
     }
+
+    public APIResponse<EmployeeResponse> getEmployeeById(Long id) {
+        Employee employee = employeeRepository.findById(id).
+                orElseThrow(() -> new RuntimeException("Employee not found"));
+        return new APIResponse<>(toEmployeeResponse(employee), "Employee retrieved successfully", true);
+    }
+
+    public APIResponse<EmployeeResponse> getEmployeeByChatID(String chatID) {
+        Employee employee = employeeRepository.findByChatID(chatID).
+                orElseThrow(() -> new RuntimeException("Employee not found"));
+
+        return new APIResponse<>(toEmployeeResponse(employee), "Employee retrieved successfully", true);
+    }
+
 
     public Employee createEmployee(CreateEmployeeDTO employeeDTO) {
         Employee employee = new Employee();
@@ -36,31 +53,66 @@ public class EmployeeService {
         return employeeRepository.save(employee);
     }
 
-//    public List<Employee> getAllEmployees()
-//    {
-//        List<Employee> all = EmployeeRepository.findAll();
-//        return all;
-//    }
+    public APIResponse<EmployeeResponse> updateEmployee(Long id, CreateEmployeeDTO employeeDTO) {
+        Employee employee = employeeRepository.findById(id).
+                orElseThrow(() -> new RuntimeException("Employee not found"));
 
-    public Employee getEmployeeById(Long id) {
-        return employeeRepository.findById(id).orElse(null);
+        if (employeeDTO.getName() != null) {
+            employee.setName(employeeDTO.getName());
+        }
+        if (employeeDTO.getDob() != null) {
+            employee.setDob(employeeDTO.getDob());
+        }
+        if (employeeDTO.getPhone() != null) {
+            employee.setPhone(employeeDTO.getPhone());
+        }
+        if (employeeDTO.getEmail() != null) {
+            employee.setEmail(employeeDTO.getEmail());
+        }
+        if (employeeDTO.getAddress() != null) {
+            employee.setAddress(employeeDTO.getAddress());
+        }
+        if (employeeDTO.getChatID() != null) {
+            employee.setChatID(employeeDTO.getChatID());
+        }
+        if (employeeDTO.getBranchID() != null) {
+            Branch branch = branchRepository.findById(employeeDTO.getBranchID()).orElseThrow(() -> new RuntimeException("Branch not found"));
+            employee.setBranch(branch);
+        }
+
+        employeeRepository.save(employee);
+        return new APIResponse<>(toEmployeeResponse(employee), "Employee updated successfully", true);
+
     }
 
-    public Employee getEmployeeByUsername(String chatID) {
-        return employeeRepository.findByChatID(chatID).orElse(null);
-    }
-
-    public Employee updateEmployee(Employee employee) {
-        return employeeRepository.save(employee);
-    }
-
-    public void deleteEmployee(Employee employee) {
+    public APIResponse<EmployeeResponse> deleteEmployee(Employee employee) {
+        if (!employeeRepository.existsById(employee.getEmployeeID())) {
+            return new APIResponse<>(null, "Employee not found", false);
+        }
         employeeRepository.delete(employee);
+        return new APIResponse<>(null, "Employee deleted successfully", true);
     }
 
-    public void deleteEmployeeById(Long id) {
+    public APIResponse<EmployeeResponse> deleteEmployeeById(Long id) {
+        if (!employeeRepository.existsById(id)) {
+            return new APIResponse<>(null, "Employee not found", false);
+        }
+
         employeeRepository.deleteById(id);
+        return new APIResponse<>(null, "Employee deleted successfully", true);
     }
 
+    public EmployeeResponse toEmployeeResponse(Employee employee) {
+       EmployeeResponse employeeResponse = new EmployeeResponse();
+       employeeResponse.setEmployeeID(employee.getEmployeeID());
+       employeeResponse.setName(employee.getName());
+       employeeResponse.setDob(employee.getDob());
+       employeeResponse.setPhone(employee.getPhone());
+       employeeResponse.setEmail(employee.getEmail());
+       employeeResponse.setAddress(employee.getAddress());
+       employeeResponse.setChatID(employee.getChatID());
+       employeeResponse.setBranchID(employee.getBranch().getBranchID());
+       return employeeResponse;
+    }
 
 }
